src/test/java/seedu/letsgethired/logic/commands/NoteCommandTest.java
package seedu.letsgethired.logic.commands;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

import static seedu.letsgethired.logic.commands.CommandTestUtil.*;
import static seedu.letsgethired.logic.commands.NoteCommand.MESSAGE_ARGUMENTS;
import static seedu.letsgethired.testutil.TypicalIndexes.INDEX_FIRST_APPLICATION;
import static seedu.letsgethired.testutil.TypicalIndexes.INDEX_SECOND_APPLICATION;
import static seedu.letsgethired.testutil.TypicalInternApplications.getTypicalInternTracker;

import org.junit.jupiter.api.Test;

import seedu.letsgethired.model.Model;
import seedu.letsgethired.model.ModelManager;
import seedu.letsgethired.model.UserPrefs;

/**
 * Contains integration tests (interaction with the Model) and unit tests for NoteCommand.
 */
public class NoteCommandTest {

    private Model model = new ModelManager(getTypicalInternTracker(), new UserPrefs());

    @Test
    public void execute() {
        final String note = "Some note";

        assertCommandFailure(new NoteCommand(INDEX_FIRST_APPLICATION, note), model,
                String.format(MESSAGE_ARGUMENTS, INDEX_FIRST_APPLICATION.getOneBased(), note));
    }

    @Test
    public void equals() {
        final NoteCommand standardCommand = new NoteCommand(INDEX_FIRST_APPLICATION, VALID_NOTE_JANE_STREET);

        // same values -> returns true
        NoteCommand commandWithSameValues = new NoteCommand(INDEX_FIRST_APPLICATION, VALID_NOTE_JANE_STREET);
        assertTrue(standardCommand.equals(commandWithSameValues));

        // same object -> returns true
        assertTrue(standardCommand.equals(standardCommand));

        // null -> returns false
        assertFalse(standardCommand.equals(null));

        // different types -> returns false
        assertFalse(standardCommand.equals(new ClearCommand()));

        // different index -> returns false
        assertFalse(standardCommand.equals(new NoteCommand(INDEX_SECOND_APPLICATION, VALID_NOTE_JANE_STREET)));

        // different note -> returns false
        assertFalse(standardCommand.equals(new NoteCommand(INDEX_FIRST_APPLICATION, VALID_COMPANY_BYTEDANCE)));
    }
}